refactor: Move TV show details entities and mappers
