+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/CastEntity.kt
@@ -1,4 +1,4 @@
-package com.london.domain.entity
+package com.london.domain.entity.tvshowdetails
 
 data class CastEntity(
     val cast: List<CastMemberEntity>,
