+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/TvShowDetailsEntity.kt
@@ -1,4 +1,4 @@
-package com.london.domain.entity
+package com.london.domain.entity.tvshowdetails
 
 data class TvShowDetailsEntity(
     val adult: Boolean,
