+++ domain/src/main/java/com/london/domain/repository/SearchRepository.kt
@@ -3,7 +3,6 @@ package com.london.domain.repository
 import com.london.domain.entity.Actor
 import com.london.domain.entity.Movie
 import com.london.domain.entity.TvShow
-import com.london.domain.entity.TvShowDetailsEntity
 
 interface SearchRepository {
     suspend fun searchForMovies(name: String, language: String): List<Movie>
