+++ presentation/src/main/java/com/london/presentation/screen/details/tvshow/tvshowdetails/TvShowDetailsUiState.kt
@@ -1,15 +1,15 @@
 package com.london.presentation.screen.details.tvshow.tvshowdetails
 
-import com.london.domain.entity.CastEntity
-import com.london.domain.entity.ImageItemEntity
-import com.london.domain.entity.TvShowCreatorEntity
-import com.london.domain.entity.TvShowEpisodeEntity
-import com.london.domain.entity.TvShowGenreEntity
-import com.london.domain.entity.TvShowNetworkEntity
-import com.london.domain.entity.TvShowProductionCompanyEntity
-import com.london.domain.entity.TvShowProductionCountryEntity
-import com.london.domain.entity.TvShowSeasonEntity
-import com.london.domain.entity.TvShowSpokenLanguageEntity
+import com.london.domain.entity.tvshowdetails.CastEntity
+import com.london.domain.entity.tvshowdetails.ImageItemEntity
+import com.london.domain.entity.tvshowdetails.TvShowCreatorEntity
+import com.london.domain.entity.tvshowdetails.TvShowEpisodeEntity
+import com.london.domain.entity.tvshowdetails.TvShowGenreEntity
+import com.london.domain.entity.tvshowdetails.TvShowNetworkEntity
+import com.london.domain.entity.tvshowdetails.TvShowProductionCompanyEntity
+import com.london.domain.entity.tvshowdetails.TvShowProductionCountryEntity
+import com.london.domain.entity.tvshowdetails.TvShowSeasonEntity
+import com.london.domain.entity.tvshowdetails.TvShowSpokenLanguageEntity
 
 data class TvShowDetailsUiState(
     val tvImages: List<ImageItemEntity>? = listOf(),
